Feed_Service/requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
dnspython==2.7.0
kafka-python==2.0.6
mongoengine==0.29.1
pymongo==4.11.2
sqlparse==0.5.3
tzdata==2025.1
